fix: move CreateTagDimensionBodySchema to schemas so msw never imports hook modules
